updated typing and exports and fixed demo types in dist. Removed unused typing. Still cannot figure out how to in InstanceType from APlayer component
