+++ demo/src/App.vue
@@ -5,6 +5,7 @@
     <h3>🍰 A beautiful HTML5 music player for Vue.js.</h3>
     <div class="aplayer-wrap">
       <a-player
+        ref="player0"
         :fixed="aplayer1.fixed"
         :lrcType="aplayer1.lrcType"
         :listMaxHeight="aplayer1.listMaxHeight"
@@ -30,10 +31,13 @@
 import './App.scss';
 import { onMounted, ref } from 'vue';
 import Hls from 'hls.js';
-import { type Options as APlayerOptions, type Audio as APlayerAudio } from '../../src/types';
 import { sleep } from '../../src/utils';
 import * as music from './assets/music/data.json';
 import '../../src/assets/style/vue-aplayer.scss'
+import { APlayer } from '../../src/index';
+import type { Options as APlayerOptions, Audio as APlayerAudio } from '../../src/index';
+
+const player0 = ref<InstanceType<typeof APlayer> | null>(null);
 
 const aplayer0 = ref<APlayerOptions>({
   fixed: true,
