+++ src/Components/Controller.vue
@@ -75,8 +75,8 @@ import Progress from './Progress.vue';
 import Button from './Button.vue';
 import VueTouch from '@/vue-touch/VueTouch.vue';
 import { inject, ref, type Ref, type ComputedRef, computed } from 'vue';
-import type { LoopMode, Media, Options, OrderMode } from '@/types/aplayer';
-import type { LrcType } from '@/types/aplayer';
+import type { LoopMode, Media, Options, OrderMode } from '@/types';
+import type { LrcType } from '@/types';
 import Play from './icons/Play.vue';
 import Pause from './icons/Pause.vue';
 import Skip from './icons/Skip.vue';
