+++ src/Components/Progress.vue
@@ -19,7 +19,7 @@
 
 <script setup lang="ts">
 import VueTouch from '@/vue-touch/VueTouch.vue';
-import type { Options } from '@/types/aplayer';
+import type { Options } from '@/types';
 import { inject, ref, computed, type Ref } from 'vue';
 import Loading from './icons/Loading.vue';
