+++ src/Components/index.ts
@@ -1,0 +1,1 @@
+export { default as APlayer } from "./APlayer.vue";
