+++ src/vue-touch/index.ts
@@ -1,0 +1,1 @@
+export { default as VueTouch } from "./VueTouch.vue";
